keep <nick/> for presence subscription stanzas
when the contact is offline, the presence subscription stanza will be
stored and resend when the contact is online.
but it doesn't store the original packet; only part of the stanza will
be stored and others will be lost.
this commit doesn't address this issue entirely; it just stores one more
element - the <nick/> element

and it passes the original packet to the handler of the
"roster_in_subscription" hook which makes it possible for the handler to
store the original packet

one more thing, we will only supports odbc, so mod_roster and
mod_shared_roster_ldap are not maintained